backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-CORS==5.0.0
Flask-Mail==0.10.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
celery==5.4.0
redis==5.2.0
Pillow==11.0.0
marshmallow==3.23.2
flask-marshmallow==1.2.1
marshmallow-sqlalchemy==0.30.0
gunicorn==23.0.0
requests==2.32.3
google-auth==2.40.3
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
bcrypt==4.2.1
google-generativeai==0.8.3
python-multipart==0.0.9
